refactor: replace VideoBar with StreamInfoBar in video components for improved layout and consistency
